module3-task1 | fix: Header и Footer вынесены из компонента Container, добавлен компонент ExtraContainer, удален компонент LoginHeader, изменена структура в папке components
