+++ src/components/catalog/card/card.tsx
@@ -5,7 +5,7 @@ export default function Card() {
         <span>Premium</span>
       </div>
       <div className="cities__image-wrapper place-card__image-wrapper">
-        <a href="#">
+        <a href="2486667-six-cities-15/src/components/catalog/card#">
           <img className="place-card__image" src="img/apartment-01.jpg" width="260" height="200" alt="Place image" />
         </a>
       </div>
@@ -29,7 +29,7 @@ export default function Card() {
           </div>
         </div>
         <h2 className="place-card__name">
-          <a href="#">Beautiful &amp; luxurious apartment at great location</a>
+          <a href="2486667-six-cities-15/src/components/catalog/card#">Beautiful &amp; luxurious apartment at great location</a>
         </h2>
         <p className="place-card__type">Apartment</p>
       </div>
