+++ src/components/common/container/container.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from 'react';
+
+type ContainerProps = {
+  children: ReactNode;
+  extraClass?: string;
+}
+
+export default function Container({ children, extraClass }: ContainerProps): JSX.Element {
+  return (
+    <main className={`page__main ${extraClass ? extraClass : ''}`}>
+      {children}
+    </main>
+  );
+}
