+++ src/components/common/header/header.tsx
@@ -11,7 +11,7 @@ export default function Header(): JSX.Element {
           <nav className="header__nav">
             <ul className="header__nav-list">
               <li className="header__nav-item user">
-                <a className="header__nav-link header__nav-link--profile" href="#">
+                <a className="header__nav-link header__nav-link--profile" href="2486667-six-cities-15/src/components/common/header#">
                   <div className="header__avatar-wrapper user__avatar-wrapper">
                   </div>
                   <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
@@ -19,7 +19,7 @@ export default function Header(): JSX.Element {
                 </a>
               </li>
               <li className="header__nav-item">
-                <a className="header__nav-link" href="#">
+                <a className="header__nav-link" href="2486667-six-cities-15/src/components/common/header#">
                   <span className="header__signout">Sign out</span>
                 </a>
               </li>
