--- src/components/header-login/header-login.tsx
@@ -1,15 +1,0 @@
-export default function HeaderLogin(): JSX.Element {
-  return (
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <a className="header__logo-link" href="main.html">
-              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-            </a>
-          </div>
-        </div>
-      </div>
-    </header>
-  );
-}
