+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/app/app';
+import App from './components/common/app/app';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
