+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,8 +1,8 @@
-import Container from '../../components/container/container';
+import Container from '../../components/common/container/container';
 
 export default function FavoritesPage(): JSX.Element {
   return (
-    <Container mainClass="page__main--favorites" isFooter footerClass="container">
+    <Container extraClass="page__main--favorites">
       <div className="page__favorites-container container">
         <section className="favorites">
           <h1 className="favorites__title">Saved listing</h1>
